feat: better atom interni, fix issue with missing required keys (#14)
In order to safely use String.to_existing_atom we need to intern the atoms used in model. This adds an @on_load hook to intern the atoms in the module from the typespec. Since all models have a typespec, this will intern and ensure every atom is available to parse to. I also tried to ensure that even if we don't parse an atom properly - we won't fail - since most of the spec is probably still correct.

Also fixes an issue where we have some "required" fields that aren't in the specification at all.

Resolves #12