feat: improve generated documentation output (#20)
Previously, the generated output was too deeply nested, resulting in it
being treated as a code block by iex and hexdocs.

To fix this, the `as_heredoc` function has been updated to take a `"""`
delimiter, which ensures that the heredoc syntax is used, instead of
manually generating the whitespace.

In addition to this, required arguments are now only displayed in the
`@doc` block if there are any, otherwise this section is omitted. Where
an argument refers to an elixir module, it is now fully qualified, so
that it generates a link to the module.

The "Sends events" section now uses the same formatting as the
arguments, using a `###` for the heading, and backticks to ensure the
code name is represented correctly. This uses a different technique, as
the events are part of the function description, and not separated out
like the arguments.