fix(ExStreamClient.JSON): only transform existing keys to atoms
Previously, all map keys were converted to atoms, and if the atom did
not exist, the ArgumentError would be rescued and an entry logged.

However, in some circumstances, the key should be treated as a string.
For example, when creating users, the response is:

```
"7332714135366402048" => %ExStreamClient.Model.FullUserResponse{
...
  updated_at: 1750070497005033000
},
"7332714224709271552" => %ExStreamClient.Model.FullUserResponse{
...
  updated_at: 1750070497005033000
}
```

In this case, the map keys are strings that will not exist in the atoms
table, as they represent use ids. To fix this, a list of valid keys is
now determined from the struct and used when checking which keys to
transform to atoms.