fix: parse datetime objects to DateTime (#36)
Noticed a few things while looking at response payloads for events: 

- DateTimes in events are always in ISO8601 (strings) format for items
described as datetimes (see here:
https://getstream.io/chat/docs/node/webhook_events/)
- API Requests (e.g. `query_channels`) encode datetimes as nanosecond
Unix timestamps.
- The specification denotes datetimes as integers (Unix timestamps) but
again we have this mismatch (datetimes in events denoted in the model as
numbers, but actually strings).

Given that we have the spec, we can actually modify this to:

- On parsing, normalize always to `DateTime.t()` at runtime and
alternatively fallback to the string / int value if that parsing fails
- Also include the other types (string / int) if the value is actually
in a request from the client TO Stream, in which case we may be encoding
differently to get the request to stream. For example, the truncate
request expects the datetime as a "number" in the spec, but happily
takes a `DateTime` encoded with `Jason.encode`, which encodes to ISO8601