feat: move response decoding to HTTP client (#38)
Resolves an underlying issue with the Stream Chat API where response
errors are gzip encoded, but provided with an incorrect content type of
`text/plain`. The result was that error messages in production above a
limit (1024 bytes?) would be incorrectly parsed, because the content was
gzipped and then not decoded as json properly.

In order to fix, the error handling was moved to `append_request_steps`
instead of `prepend_request_steps` so that `Req` could properly
`decompress` in the default response steps, THEN be processed by the
http client steps.

However, this broke the ordering of the original `decode` step that was
added as `append_response_steps` in the operations modules. Because this
was called first, it was BEFORE the `decode_error_json` step, which was
previously in `prepend`. So, as a side effect, we needed to move
response handling into the http client. This moves (and ultimately
lowers the total delivered code bytes) response handling to the http
client. The client now optionally accepts a `response_handlers` map that
contains response codes and modules for nested decoding, and will apply
those handlers if the response code is known.